the activity id must be in the activities collection before changing its state
prior to the recent refactor / assert additions, this worked because setting
the state would, as a side-effect, add it to the activities list. obviously this
is not desirable in the case of, say, starting an activity so that change makes sense
but it made it impossible to add new activities without hitting the assert in
setActivityState.

this library really ought to have unit tests. :/