Loading the linking options only on request
Testing whether selected files are linked to the activity or not
is slow. We are doing it now only on request, in a separate thread
so not to block the UI for the user.

Even more, if a large number of files has been selected, we'll
show all options, without checking which ones are already linked.