Adding the option not to keep the stats for specific activities
We already have the global options for the user to specify
which applications should be tracked, and for how long the
statistics should be kept, or to turn the tracking off completely.

Now, it is possible to turn off tracking for specific activities.
This will provide a feature similar to the private browsing mode
commonly found in modern web browsers.

CHANGELOG: Adding the option not to track the user when in specific activities (similar to the 'private browsing' mode in a web browser)