Replacing QDBusInterface usage with the interface generated by qdbusxml2cpp
Summary:
This patch replaces QDBusInterface usage for checking KAMD service version with the interface generated by qdbusxml2cpp.

The reason is that QDBusInterface has blocking introspection calls in its constructor (!?) which can block for up to 25 seconds. According to Thiago (I've found this bug from the Harmattan era) it is not a bug. Or it is a really low priority one, and we should all use just interfaces generated with qdbusxml2cpp.

Reviewers: graesslin, davidedmundson

Reviewed By: davidedmundson

Subscribers: davidedmundson

Projects: #kactivities

Differential Revision: https://phabricator.kde.org/D544