Fix crash in activities context menu
Currently we call deleteLater() from inside ::run which is running in a
different thread than the receiving object.
(QThread objects live in the thread that created them, not in the thread
they create)

This patch causes deleteLater to be run in the right thread.

QCoreApplication::postEvent is thread safe but it needs to be in the
right thread to work out the correct event loop level for deferred
delete events.

BUG: 351485
REVIEW: 126955