Fix race that starts kactivitymanagerd multiple times
Summary:
During plasma startup we have multiple processes using kactivities at
once. The current code looks to see if the bus is active then launched
the executable.

If this happens at once, we end up loading, doing all the
initialisation only for all-but-one of the services to fail to register
the name
and then quit.

By using DBus activation, we get that global lock in a very safe way.

This requires a change (a .service file) on the daemon, so a fallback is
provided for 5.11 support

Test Plan:
Had just this change.
Killed kactivitymanagerd
Restarted plasmashell. It launched
Made workspace change. Had activities.

Subscribers: #frameworks

Tags: #frameworks

Differential Revision: https://phabricator.kde.org/D8732