Drop most of ResourceInstance
Barely any app uses the full version of this. Most apps that do integrate with activities resources
use the static notifyAccessed function.

The full variant also is effectively unused in the backend. It was used for the share-like-connect feature, but that was dropped.

The focus tracking part also doesn't work as-is on Wayland due to relying on global window ids and kamd being able to monitor windows.
This is solvable in principe, but not necessarily without API changes or changing the mechansim completely.

Given the lack of usefulness of the full variant remove that and port the few existing consumers to the slimmer static function